test/smoke: split up and rename files
