+++ test/smoke/testcluster_test.go
@@ -1,0 +1,10 @@
+//+build smoke
+
+package smoke
+
+type TestCluster struct {
+	BaseDir      string
+	RemoteGDMDir string
+	Count        int
+	Instances    []*Instance
+}
